eth_scalper/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.11.0
python-telegram-bot>=20.0
flask>=3.0.0
